requirements.txt
apache-airflow-providers-common-ai[sql,mcp,parquet,avro]==0.1.0
apache-airflow-providers-sqlite==4.3.2
pydantic-ai-slim[openai,anthropic,mcp]==1.85.1
datafusion==53.0.0
httpx==0.28.1
